server edits

+++ src/main/java/com/example/graphfunctionsapp/Function3D.java
@@ -1,8 +1,8 @@
 package com.example.graphfunctionsapp;
 
 public class Function3D {
-    public float compute(float x, float y, float param1, float param2) {
+    public double compute(double x, double y, double param1, double param2) {
         // Здесь можно изменить формулу на любую другую функцию
-        return (float) (Math.sin(x * param1) * Math.cos(y * param2));
+        return (Math.sin(x * param1) * Math.cos(y * param2));
     }
 }
